# docs/specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LoonFS HTTP API",
    "description": "Static OpenAPI document for the LoonFS v0 HTTP API.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.2.1"
  },
  "paths": {
...
    "/v0/admin/namespaces/{namespace_id}/grep/index": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Read the grep index's lifecycle",
        "description": "Reports where the namespace's grep index is: `disabled`, `backfilling` with the sequence it walks toward and how far the walk got, or `active` with the watermark it has built through. One grep root read, no side effects. A namespace that never enabled the index reads as `disabled`. Requires this deployment to maintain the grep index.",
        "operationId": "get_grep_index_status",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The index's lifecycle and bookkeeping",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrepIndexStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The backing store rejected its configured credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "500": {
            "description": "The grep index is corrupt or its backing store is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "This deployment does not maintain the grep index",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/admin/namespaces/{namespace_id}/grep/index/disable": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Disable the grep index",
        "description": "Disables the namespace's grep root and clears its segment references with one durable compare-and-swap; index maintenance stops on its own once a step reads the disabled root. Explicit grep garbage collection later reclaims the segments. Idempotent. Requires this deployment to maintain the grep index.",
        "operationId": "disable_grep_index",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grep root disabled or already disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrepIndexStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The backing store rejected its configured credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "409": {
            "description": "Lost a grep root-pointer publication race; retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "The grep index is corrupt or its backing store is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "This deployment does not maintain the grep index",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/admin/namespaces/{namespace_id}/grep/index/enable": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Enable the grep index",
        "description": "Enables the namespace's grep root and asks this deployment's maintenance runner for the backfill's first step. The response reports the lifecycle and bookkeeping read after the transition: a fresh enable is `backfilling` with the sequence its checkpoint captured, while an already-enabled namespace answers with its current status. Idempotent. Requires this deployment to maintain the grep index.",
        "operationId": "enable_grep_index",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grep root enabled or already enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrepIndexStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The backing store rejected its configured credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "409": {
            "description": "Lost a grep root-pointer publication race; retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "The grep index is corrupt or its backing store is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "This deployment does not maintain the grep index",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/admin/namespaces/{namespace_id}/grep/index/gc": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Collect grep-index garbage",
        "description": "Runs one explicit garbage-collection pass over only this namespace's grep-owned extension keyspace. A tombstoned or absent namespace has aged extension state reaped; no grep garbage collection runs implicitly. `max_objects` bounds the reads the pass spends and returns a `next_cursor` when keys remain; resuming re-reads liveness and the grep root, so a cursor only skips enumeration. Requires this deployment to maintain the grep index.",
        "operationId": "gc_grep_index",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/GrepGcRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Namespace grep garbage collection completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrepGcResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid budget or cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The backing store rejected its configured credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "The grep index is corrupt or its backing store is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "This deployment does not maintain the grep index",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v0/namespaces/{namespace_id}/query/grep": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Content search",
        "description": "Searches file content with a regular expression, accelerated by the namespace's grep index. Matches are verified against the real pattern and returned in ascending `(inode_id, byte_offset)` order; revisions committed after the index watermark are scanned exhaustively unless `allow_stale` skips them. Requires this deployment to serve grep and the namespace to carry a materialized active grep root.",
        "operationId": "grep",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "One page of matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrepResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pattern, cursor, or an unindexable pattern without allow_scan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The backing store rejected its configured credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "410": {
            "description": "Namespace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "The grep index is corrupt or its backing store is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "This deployment does not serve grep queries, the grep index is not enabled, or its backfill has not completed on this namespace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "The index trails the head past the scan budget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "HTTP error body used by LoonFS APIs.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable reason from the [`ErrorCode`](crate::ErrorCode)\nregistry.\n\nCarried as a string so clients keep working when a newer server\nintroduces a code they do not know; use\n[`ErrorCode::parse`](crate::ErrorCode::parse) for typed access."
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorDetails",
                "description": "Structured context for the code, present when the failure carries\nmachine-usable identity (API spec, \"Standard error contract\"). Boxed\nso the rare detailed error does not widen every error-carrying result."
              }
            ]
          },
          "feature": {
            "type": [
              "string",
              "null"
            ],
            "description": "For `not_supported` errors, the capability-document feature key the\nclient should reconcile against."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          },
          "param": {
            "type": [
              "string",
              "null"
            ],
            "description": "Identifies the invalid input. Body fields use JSON Pointer paths;\nquery and path parameters use their names; CLI errors use the flag or\nargument as written."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation id the server assigned to the failed request; the same\nvalue is sent as the `x-request-id` response header."
          }
        }
      },
...
      "GrepIndexStatusResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/GrepIndexLifecycle",
            "description": "Where the index is in its lifecycle."
          },
          {
            "type": "object",
            "required": [
              "namespace_id",
              "next_run_ordinal",
              "reorganize_pending"
            ],
            "properties": {
              "namespace_id": {
                "$ref": "#/components/schemas/NamespaceId",
                "description": "Namespace the status describes."
              },
              "next_run_ordinal": {
                "type": "integer",
                "format": "int64",
                "description": "Next logical run ordinal the index will allocate.",
                "maximum": 9007199254740991,
                "minimum": 0
              },
              "reorganize_pending": {
                "type": "boolean",
                "description": "True while a partitioned segment reorganization is in progress."
              }
            }
          }
        ],
        "description": "The namespace's grep-index lifecycle and its cheap bookkeeping (admin\nplane)."
      },
...
    },
...
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The deployment's `auth_token`, sent as `Authorization: Bearer <token>`. A server configured without a token accepts every request; one configured with a token answers 401 `unauthorized` without it."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}